+++ internal/protobuf/seekers.go
@@ -76,3 +76,19 @@ func GetLENFieldBounds(buf []byte, num protowire.Number) (FieldBounds, error) {
 
 	return ParseLENFieldBounds(buf, off, tagLn, num, typ)
 }
+
+// GetUint64Field seeks uint64 field in buf by number and parses it. If field is
+// missing, no error is returned.
+//
+// Message should have ascending field order, otherwise error returns.
+//
+// If there is an error, its text contains num.
+func GetUint64Field(buf []byte, num protowire.Number) (uint64, error) {
+	off, tagLn, typ, err := SeekFieldByNumber(buf, num)
+	if err != nil {
+		return 0, err
+	}
+
+	u, _, err := ParseUint64Field(buf[off+tagLn:], num, typ)
+	return u, err
+}