+++ pkg/local_object_storage/blobstor/internal/storagetest/delete.go
@@ -35,7 +35,7 @@ func TestDelete(t *testing.T, cons Constructor, minSize, maxSize uint64) {
 				require.ErrorAs(t, err, new(apistatus.ObjectNotFound))
 			})
 			t.Run("getrange fail", func(t *testing.T) {
-				_, err := s.GetRangeStream(oidtest.Address(), 0, 1)
+				_, _, err := s.GetRangeStream(oidtest.Address(), 0, 1)
 				require.ErrorAs(t, err, new(apistatus.ObjectNotFound))
 			})
 		})