+++ pkg/local_object_storage/shard/ec.go
@@ -108,6 +108,21 @@ func (s *Shard) getECPartFunc(cnr cid.ID, parent oid.ID, pi iec.PartInfo, writeC
 	return nil
 }
 
+// ReadECPartRange is a buffered alternative for [Shard.GetECPartRange]
+// similar to [Shard.ReadHeader].
+func (s *Shard) ReadECPartRange(cnr cid.ID, parent oid.ID, pi iec.PartInfo, off, ln uint64, buf []byte) (io.ReadCloser, error) {
+	var stream io.ReadCloser
+	return stream, s.getECPartRangeFunc(cnr, parent, pi, off, ln, func(writeCache writecache.Cache, addr oid.Address, off, ln uint64) error {
+		var err error
+		stream, err = writeCache.ReadPayloadRange(addr, off, ln, buf)
+		return err
+	}, func(blobStorage common.Storage, addr oid.Address, off, ln uint64) error {
+		var err error
+		stream, err = blobStorage.ReadPayloadRange(addr, off, ln, buf)
+		return err
+	})
+}
+
 // GetECPartRange looks up for object that carries EC part produced within cnr
 // for parent object and indexed by pi in the underlying metabase, checks its
 // availability and reads it from the underlying BLOB storage. Returns full
@@ -134,11 +149,28 @@ func (s *Shard) getECPartFunc(cnr cid.ID, parent oid.ID, pi iec.PartInfo, writeC
 //
 // If the range is out of payload bounds, GetECPartRange returns
 // [apistatus.ErrObjectOutOfRange].
-func (s *Shard) GetECPartRange(cnr cid.ID, parent oid.ID, pi iec.PartInfo, off, ln int64) (uint64, io.ReadCloser, error) {
+func (s *Shard) GetECPartRange(cnr cid.ID, parent oid.ID, pi iec.PartInfo, off, ln uint64) (uint64, io.ReadCloser, error) {
+	var pldLen uint64
+	var stream io.ReadCloser
+	return pldLen, stream, s.getECPartRangeFunc(cnr, parent, pi, off, ln, func(writeCache writecache.Cache, addr oid.Address, off, ln uint64) error {
+		var err error
+		pldLen, stream, err = writeCache.GetRangeStream(addr, off, ln)
+		return err
+	}, func(blobStorage common.Storage, addr oid.Address, off, ln uint64) error {
+		var err error
+		pldLen, stream, err = blobStorage.GetRangeStream(addr, off, ln)
+		return err
+	})
+}
+
+func (s *Shard) getECPartRangeFunc(cnr cid.ID, parent oid.ID, pi iec.PartInfo, off, ln uint64,
+	writeCacheFn func(writecache.Cache, oid.Address, uint64, uint64) error,
+	blobStorageFn func(common.Storage, oid.Address, uint64, uint64) error,
+) error {
 	partID, pldLen, err := s.metaBaseIface.ResolveECPartWithPayloadLen(cnr, parent, pi)
 	if err != nil {
 		if !errors.As(err, (*ierrors.ObjectID)(&partID)) {
-			return 0, nil, fmt.Errorf("resolve part ID and payload len in metabase: %w", err)
+			return fmt.Errorf("resolve part ID and payload len in metabase: %w", err)
 		}
 
 		s.log.Warn("EC part ID returned from metabase with error",
@@ -146,19 +178,19 @@ func (s *Shard) GetECPartRange(cnr cid.ID, parent oid.ID, pi iec.PartInfo, off, 
 	} else {
 		if ln == 0 && off == 0 {
 			if pldLen == 0 {
-				return 0, nil, nil
+				return nil
 			}
-		} else if uint64(off) >= pldLen || pldLen-uint64(off) < uint64(ln) {
-			return 0, nil, apistatus.ErrObjectOutOfRange
+		} else if off >= pldLen || pldLen-off < ln {
+			return apistatus.ErrObjectOutOfRange
 		}
 	}
 
-	pldLen, rc, err := s.GetRangeStream(cnr, partID, off, ln)
+	err = s.getRangeStreamFunc(cnr, partID, off, ln, writeCacheFn, blobStorageFn)
 	if err != nil {
-		return 0, nil, fmt.Errorf("get range by ID %w: %w", ierrors.ObjectID(partID), err)
+		return fmt.Errorf("get range by ID %w: %w", ierrors.ObjectID(partID), err)
 	}
 
-	return pldLen, rc, nil
+	return nil
 }
 
 // ReadECPartHeader is a buffered alternative for [Shard.HeadECPart]