+++ pkg/local_object_storage/writecache/get_test.go
@@ -79,14 +79,14 @@ func TestCache_GetRangeStream(t *testing.T) {
 
 	addr := obj.Address()
 
-	_, err := c.GetRangeStream(addr, 0, 0)
+	_, _, err := c.GetRangeStream(addr, 0, 0)
 	require.ErrorIs(t, err, apistatus.ErrObjectNotFound)
-	_, err = c.GetRangeStream(addr, 1, pldLen-1)
+	_, _, err = c.GetRangeStream(addr, 1, pldLen-1)
 	require.ErrorIs(t, err, apistatus.ErrObjectNotFound)
 
 	require.NoError(t, c.Put(addr, &obj, obj.Marshal()))
 
-	_, err = c.GetRangeStream(addr, 1, 0)
+	_, _, err = c.GetRangeStream(addr, 1, 0)
 	require.EqualError(t, err, "invalid range off=1,ln=0")
 
 	for _, tc := range []struct{ off, ln uint64 }{
@@ -95,8 +95,9 @@ func TestCache_GetRangeStream(t *testing.T) {
 		{off: 1, ln: pldLen - 1},
 		{off: pldLen - 1, ln: 1},
 	} {
-		stream, err := c.GetRangeStream(addr, tc.off, tc.ln)
+		gotPldLen, stream, err := c.GetRangeStream(addr, tc.off, tc.ln)
 		require.NoError(t, err, tc)
+		require.EqualValues(t, pldLen, gotPldLen)
 
 		b, err := io.ReadAll(stream)
 		require.NoError(t, err)
@@ -115,15 +116,15 @@ func TestCache_GetRangeStream(t *testing.T) {
 		{off: 1, ln: pldLen},
 		{off: pldLen - 1, ln: 2},
 	} {
-		_, err := c.GetRangeStream(addr, tc.off, tc.ln)
+		_, _, err := c.GetRangeStream(addr, tc.off, tc.ln)
 		require.ErrorIs(t, err, apistatus.ErrObjectOutOfRange)
 	}
 
 	require.NoError(t, c.Delete(addr))
 
-	_, err = c.GetRangeStream(addr, 0, 0)
+	_, _, err = c.GetRangeStream(addr, 0, 0)
 	require.ErrorIs(t, err, apistatus.ErrObjectNotFound)
-	_, err = c.GetRangeStream(addr, 1, pldLen-1)
+	_, _, err = c.GetRangeStream(addr, 1, pldLen-1)
 	require.ErrorIs(t, err, apistatus.ErrObjectNotFound)
 }